Consolidate app install/management helpers and trim the public surface
Merge lib/commands/app-install.ts into lib/commands/helpers/app.ts, and
relocate checkAutInstallationState/installOtherApps there too, keeping
app-management.ts focused on the mobile: app lifecycle commands. Move the
AutInstallationState(Options) types out of driver.ts into commands/types.ts.

Un-export verifyApplicationPlatform, installOtherApps, downloadIpa and
unzipApp now that they're only called internally, and drop unzipFile/
unzipStream/UnzipInfo from the helpers barrel since they're only used
within helpers/app.ts and by one test file, which now imports them
directly.
